+++ .vscode/launch.json
@@ -106,7 +106,7 @@
             "name": "standalone_ui/editor_sidebar + experiments",
             "request": "launch",
             "type": "dart",
-            "program": "devtools_app/test/test_infra/scenes/standalone_ui/editor_sidebar.stager_app.g.dart",
+            "program": "packages/devtools_app/test/test_infra/scenes/standalone_ui/editor_sidebar.stager_app.g.dart",
             "preLaunchTask": "Start DTD on Port 8500",
             "args": [
                 "--dart-define=enable_experiments=true"
@@ -116,7 +116,7 @@
             "name": "standalone_ui/property_editor_sidebar",
             "request": "launch",
             "type": "dart",
-            "program": "devtools_app/test/test_infra/scenes/standalone_ui/property_editor_sidebar.stager_app.g.dart",
+            "program": "packages/devtools_app/test/test_infra/scenes/standalone_ui/property_editor_sidebar.stager_app.g.dart",
             "preLaunchTask": "Start DTD on Port 8500",
         },
         {
